requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyPDF2
huggingface_hub
